modified:   lib/connection/Connection.h 	modified:   lib/explorer/Explorer.cpp 	modified:   lib/explorer/Explorer.h 	modified:   lib/navigator/Navigator.cpp 	modified:   lib/ultrasonic/Ultrasonic.h 	modified:   platformio.ini 	modified:   src/main.cpp

+++ lib/ultrasonic/Ultrasonic.h
@@ -6,7 +6,7 @@
 static const uint8_t TRIG_PIN = 17;
 static const uint8_t ECHO_PIN = 16;
 static const uint8_t MIN_DISTANCE = 2;
-static const uint8_t MAX_DISTANCE = 400;
+static const uint16_t MAX_DISTANCE = 400;
 
 class Ultrasonic
 {
